test 2

+++ bash_scripts/save_to_drive.sh
@@ -17,8 +17,9 @@ done
 source_directory="/content/test_colab/"
 destination_directory="/content/drive/MyDrive/Colab Notebooks/UNF_Fysik_Camp"
 
-for file in "$source_directory"/*.ipynb | grep -v 'THIS.ipynb'; do
-  if [ ! -f "$destination_directory/$(basename "$file")" ]; then
+for file in "$source_directory"/*.ipynb; do
+  if [ ! -f "$destination_directory/$(basename "$file")" ] && [ "${file##*/}" != "THIS.ipynb" ]; then
+    echo "Copying $file to $destination_directory"
     cp "$file" "$destination_directory"
   fi
 done
